Refactor Utils class to create ShutdownHookManager
